move file handling into dedicated sketch bloc
